refactor(test): split interview page tests into separate files
